refactor: Add a paragraph and modify some heading tags
Add a paragraph to Featured Work to meet required elements specifications as stated in the rubric for this project.
Modify the projects anchor tags, weights and colors to be more visible and apply a new style size to the h1 header tag to increase its importance in the page context for bigger viewports.