refactor: Add a horizontal rule, restyle borders and change text
Add a horizontal rule at the end of the page to keep the last image/images separated from the bottom of the screen and stylize all the three horizontal rules consistently.
Add a similar style to the article images to make them stand out more.
Add an anchor tag and the Udacity name to the paragraph text.